Tests/LibWeb/Text/expected/css/transform-non-invertible-discrete-fallback.txt

at 0.2: matrix(2, 0, 0, 2, 0, 0)
at 0.8: matrix(1, 0, 0, 1, 0, 0)
